Revert "Moved NetConfigManager to ConnectionStrings"
This reverts commit 641bd9387634f2f8ec8ee13a3b7b9d79b2ab1121.
